RX-6812: Kill the tunnel on disconnect
The tunnel is established by a java program, that does not exit by
itself when the tunnel disconnects. To work around that, a small wrapper
shell script monitors the stdout of the java program and stops it when
lines indicate that the connection has been lost. Additionally, the tini
init process will ensure that child processes are reaped correctly.